% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adstock.R
\name{adstock}
\alias{adstock}
\title{Calculate adstock (decay)}
\usage{
adstock(x, rate = 0)
}
\arguments{
\item{x}{Numeric vector to be passed through.}

\item{rate}{Decay rate to be applied to `x`. Must be between 0 and 1.}
}
\value{
A numeric vector of the same length as `x` with the adstock
  transformation applied.
}
\description{
Applies an adstock (decay) transformation to a numeric vector using
a recursive filter.
}
\examples{
# Apply 20\% decay rate
adstock(c(100, 200, 300, 150, 200), rate = 0.2)

}
