% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumlagdiff.R
\name{sumlagdiff}
\alias{sumlagdiff}
\title{Create a score that measures the absolute fluctuation
  for a numeric vector}
\usage{
sumlagdiff(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical. Should missing values (including `NaN`) be removed?}
}
\value{
A single numeric value representing the sum of absolute differences.
}
\description{
Returns the sum of absolute differences for each
value in a vector, between \emph{k} and \emph{k-1}.
}
\examples{
p <- c(3, 5, 4, 2, 1, 7)
q <- c(4, 4, 3, 3, 4, 4)
sumlagdiff(p)
sumlagdiff(q)

}
