% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adstock.R
\name{adstock}
\alias{adstock}
\title{Calculate adstock (decay)}
\usage{
adstock(x, rate = 0)
}
\arguments{
\item{x}{Numeric vector to be passed through.}

\item{rate}{Decay rate to be applied to `x`}
}
\description{
Calculate adstock (decay)
}
