% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_direction.R
\name{analyse_direction}
\alias{analyse_direction}
\title{Analyse co-movement between two numeric variables.}
\usage{
analyse_direction(x, var1, var2)
}
\description{
An explanatory note is printed as a message in the console.
}
