% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{direction}
\alias{direction}
\title{For a numeric vector of values, return its relation to the previous (lagged) value}
\usage{
direction(x)
}
\arguments{
\item{x}{Numeric vector to pass through}
}
\description{
For a numeric vector of values, return its relation to the previous (lagged) value
}
