% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_k_date.R
\name{return_k_date}
\alias{return_k_date}
\title{Return the k-th most recent or oldest date / datetime from a vector}
\usage{
return_k_date(x, k, decreasing = TRUE)
}
\arguments{
\item{x}{A vector of date-time}

\item{k}{Integer specifying the k-th value to return from the vector}

\item{decreasing}{Logical. Specifies whether to return most recent or oldest}
}
\description{
Return the k-th most recent or oldest date / datetime from a vector
}
\examples{
library(lubridate)
dates <- c(ymd("2018-01-01"), ymd("2016-01-31"), ymd("2017-01-31"))
return_k_date(dates, k = 2)
}
