% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_adstock.R
\name{reverse_adstock}
\alias{reverse_adstock}
\title{Convert adstocked values back to original values}
\usage{
reverse_adstock(x, rate = 0)
}
\arguments{
\item{x}{Numeric vector to be passed through}

\item{rate}{Adstock rate to be used (must be a positive value)}
}
\description{
Convert adstocked values back to original values
}
