% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stend_line.R
\name{stend_line}
\alias{stend_line}
\title{Start-End Line}
\usage{
stend_line(x)
}
\arguments{
\item{x}{Numeric vector to pass through. Missing values other than the start and end value
in the vector are ignored.}
}
\description{
Generates a linear vector between the start and end value of the input vector.
}
