% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_summarise.R
\name{ts_summarise}
\alias{ts_summarise}
\title{Group-summarise a time series}
\usage{
ts_summarise(x, grouping = "year", date_var = "Date", fun = "sum",
  ...)
}
\arguments{
\item{x}{Data frame to be passed through}

\item{grouping}{Method of grouping / rounding used for dates. See `lubridate::floor_date()`}

\item{date_var}{Character string specifying variable name with the date variable}

\item{fun}{Character string specifying function for summarising. Defaults to sum}

\item{...}{Additional arguments}
}
\description{
Group-summarise a time series
}
