% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcf.R
\name{xcf}
\alias{xcf}
\title{Create a cross-correlation table}
\usage{
xcf(df, x, y)
}
\arguments{
\item{x}{Numeric vector to pass through}

\item{y}{Numeric vector to pass through}
}
\description{
Returns the values of the cross-correlation correlation
}
